% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignmentProb.R
\name{alignmentProb}
\alias{alignmentProb}
\title{Computes site probabilities for a binary character model}
\usage{
alignmentProb(tree, tree.states, inactive.prob = 0.95, rate = 10^-3,
  log = TRUE)
}
\arguments{
\item{tree}{a phylo object from the ape package}

\item{tree.states}{a matrix of 0 and 1 values where each column is a species and each row is a site}

\item{inactive.prob}{The probability that state=0 @ the root of the tree}

\item{rate}{the mutational rate}

\item{log}{if TRUE state probabilies are returned in log space}
}
\value{
a vector containing the per-site probability
}
\description{
Computes site probabilities for a binary character model
}
