% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeASRunInvariants.R
\name{computeASRunInvariants}
\alias{computeASRunInvariants}
\title{Pre-computes all values needed to compute the probability of the absorbing state}
\usage{
computeASRunInvariants(tree, rate, base.freq.zero)
}
\arguments{
\item{tree}{a phylo object}

\item{rate}{a mutation rate}

\item{base.freq}{a vector, (p(0),p(1)) of the probabilities of 0/1 at the root node}

\item{log}{if TRUE returns log transition matricies}
}
\value{
list of parameter values that are constant for calculating the probability of the absorbing state accross sites
}
\description{
Pre-computes all values needed to compute the probability of the absorbing state
}
