% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTraversalTable.R
\name{createTraversalTable}
\alias{createTraversalTable}
\title{Returns the leaves of a binary tree and a post-order traversal}
\usage{
createTraversalTable(tree)
}
\arguments{
\item{tree}{a phylo object from the ape package}
}
\value{
a list containing the tree tips and a postorder traversal
}
\description{
Returns the leaves of a binary tree and a post-order traversal
}
