% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expMatList.R
\name{expMatList}
\alias{expMatList}
\title{Compute exponentiated rate matrix for all branches}
\usage{
expMatList(rate, base.freq.zero, branch.lengths, log = TRUE)
}
\arguments{
\item{rate}{a mutation rate}

\item{branch.lengths}{the branch lengths of interes}

\item{log}{if TRUE returns log transition matricies}
}
\value{
a list of transition matricies, one entry for each specified branch length
}
\description{
Compute exponentiated rate matrix for all branches
}
