% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logMinusExp}
\alias{logMinusExp}
\title{Vector logMinusExp}
\usage{
logMinusExp(x, y)
}
\arguments{
\item{x}{A numeric vector}

\item{a}{A scalar}
}
\description{
This function computes the  e^a-sum(e^x) of a scalar 'a' and vector 'x' without leaving log space
}
